+++ src/TextInputWidget.h
@@ -30,7 +30,6 @@
 
 #include "SuggestionsPopup.h"
 #include "dialogs/PreviewUploadOverlay.h"
-#include "emoji/PickButton.h"
 
 namespace dialogs {
 class PreviewUploadOverlay;
@@ -160,9 +159,6 @@ public slots:
         void focusLineEdit() { input_->setFocus(); }
         void addReply(const QString &username, const QString &msg);
 
-private slots:
-        void addSelectedEmoji(const QString &emoji);
-
 signals:
         void sendTextMessage(QString msg);
         void sendEmoteMessage(QString msg);
@@ -193,7 +189,6 @@ private:
 
         FlatButton *sendFileBtn_;
         FlatButton *sendMessageBtn_;
-        emoji::PickButton *emojiBtn_;
 
         QColor borderColor_;
 };